+++ raw.go
@@ -1,8 +1,10 @@
 package gosnowth
 
 import (
+	"bytes"
 	"context"
 	"encoding/json"
+	"fmt"
 	"io"
 	"net/http"
 	"net/url"
@@ -10,7 +12,10 @@ import (
 	"strconv"
 	"time"
 
+	flatbuffers "github.com/google/flatbuffers/go"
 	"github.com/pkg/errors"
+
+	"github.com/circonus-labs/gosnowth/fb/noit"
 )
 
 // FlatbufferContentType is the content type header for flatbuffer data.
@@ -78,25 +83,69 @@ func (sc *SnowthClient) ReadRawNumericValuesContext(ctx context.Context,
 	return r.Data, nil
 }
 
+// WriteRawResponse values represent raw IRONdb data write responses.
+type WriteRawResponse struct {
+	Errors      uint64 `json:"errors"`
+	Misdirected uint64 `json:"misdirected"`
+	Records     uint64 `json:"records"`
+	Updated     uint64 `json:"updated"`
+}
+
 // WriteRaw writes raw IRONdb data to a node.
 func (sc *SnowthClient) WriteRaw(node *SnowthNode, data io.Reader,
-	fb bool, dataPoints uint64) error {
+	fb bool, dataPoints uint64) (*WriteRawResponse, error) {
 	return sc.WriteRawContext(context.Background(), node, data, fb, dataPoints)
 }
 
 // WriteRawContext is the context aware version of WriteRaw.
 func (sc *SnowthClient) WriteRawContext(ctx context.Context, node *SnowthNode,
-	data io.Reader, fb bool, dataPoints uint64) error {
+	data io.Reader, fb bool, dataPoints uint64) (*WriteRawResponse, error) {
 
 	hdrs := http.Header{"X-Snowth-Datapoints": {strconv.FormatUint(dataPoints, 10)}}
 	if fb { // is flatbuffer?
 		hdrs["Content-Type"] = []string{FlatbufferContentType}
 	}
 
-	_, _, err := sc.do(ctx, node, "POST", "/raw", data, hdrs)
+	body, _, err := sc.do(ctx, node, "POST", "/raw", data, hdrs)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
-	return nil
+	r := &WriteRawResponse{}
+	if err := decodeJSON(body, &r); err != nil {
+		return nil, errors.Wrap(err, "unable to decode IRONdb response")
+	}
+
+	return r, nil
+}
+
+var metricListFileIdentifier = []byte("CIML")
+
+// WriteRawMetricList writes raw IRONdb data to a node with FlatBuffers.
+func (sc *SnowthClient) WriteRawMetricList(node *SnowthNode, metricList *noit.MetricListT,
+	builder *flatbuffers.Builder) (*WriteRawResponse, error) {
+	return sc.WriteRawMetricListContext(context.Background(), node, metricList, builder)
+}
+
+// WriteRawMetricListContext is the context aware version of WriteRawMetricList.
+func (sc *SnowthClient) WriteRawMetricListContext(ctx context.Context, node *SnowthNode,
+	metricList *noit.MetricListT, builder *flatbuffers.Builder) (*WriteRawResponse, error) {
+
+	if metricList == nil {
+		return nil, fmt.Errorf("metric list cannot be nil")
+	}
+	datapoints := uint64(len(metricList.Metrics))
+	if datapoints == 0 {
+		return nil, fmt.Errorf("metric list cannot be empty")
+	}
+	if builder == nil {
+		builder = flatbuffers.NewBuilder(1024)
+	} else {
+		builder.Reset()
+	}
+	offset := noit.MetricListPack(builder, metricList)
+	builder.FinishWithFileIdentifier(offset, metricListFileIdentifier)
+	reader := bytes.NewReader(builder.FinishedBytes())
+
+	return sc.WriteRawContext(ctx, node, reader, true, datapoints)
 }